+++ src/filters/BaseFilter.ts
@@ -9,36 +9,29 @@ import type {
 } from './typedefs';
 import { isWebGLPipelineState } from './typedefs';
 import { GLPrecision } from './GLProbes/GLProbe';
+import {
+  highPsourceCode,
+  identityFragmentShader,
+  vertexSource,
+} from './shaders/baseFilter';
 
-const highPsourceCode = `precision ${GLPrecision.high} float`;
-
-export type AbstractBaseFilterOptions<T> = {
-  mainParameter: string;
-  vertexSource: string;
-  fragmentSource: T;
-};
-
-export type BaseFilterOptions = AbstractBaseFilterOptions<string>;
-
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface AnyFilter
-  extends AbstractBaseFilter<string | Record<string, string>> {}
-
-export abstract class AbstractBaseFilter<T> {
+export class BaseFilter {
   /**
    * Filter type
    * @param {String} type
    * @default
    */
-  declare type: string;
+  get type(): string {
+    return this.constructor.name;
+  }
+
+  declare static defaults: Record<string, any>;
 
   /**
    * Array of attributes to send with buffers. do not modify
    * @private
    */
-  declare vertexSource: string;
-
-  declare fragmentSource: T;
+  vertexSource = vertexSource;
 
   /**
    * Name of the parameter that can be changed in the filter.
@@ -52,13 +45,17 @@ export abstract class AbstractBaseFilter<T> {
    * Constructor
    * @param {Object} [options] Options object
    */
-  constructor(
-    options: Partial<AbstractBaseFilterOptions<T>> & Record<string, any> = {}
-  ) {
-    Object.assign(this, options);
+  constructor({ ...options }: Record<string, any> = {}) {
+    Object.assign(
+      this,
+      (this.constructor as typeof BaseFilter).defaults,
+      options
+    );
   }
 
-  abstract getFragmentSource(): string;
+  protected getFragmentSource(): string {
+    return identityFragmentShader;
+  }
 
   /**
    * Compile this filter's shader program.
@@ -151,10 +148,12 @@ export abstract class AbstractBaseFilter<T> {
    * @param {WebGLShaderProgram} program The shader program from which to take uniform locations.
    * @returns {Object} A map of uniform names to uniform locations.
    */
-  abstract getUniformLocations(
+  getUniformLocations(
     gl: WebGLRenderingContext,
     program: WebGLProgram
-  ): TWebGLUniformLocationMap;
+  ): TWebGLUniformLocationMap {
+    return {};
+  }
 
   /**
    * Send attribute data from this filter to its shader program on the GPU.
@@ -220,16 +219,17 @@ export abstract class AbstractBaseFilter<T> {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   isNeutralState(options?: any): boolean {
     const main = this.mainParameter,
-      // @ts-ignore ts you are lying
-      proto = this.__proto__;
+      defaultValue = (this.constructor as typeof BaseFilter).defaults[
+        main as string
+      ];
     if (main) {
-      if (Array.isArray(proto[main]) && Array.isArray(this[main])) {
-        return proto[main].every(
-          // @ts-ignore requires some kind of dynamic type thing, or delete, or leave it ignored
-          (value: any, i: number) => value === this[main][i]
+      const thisValue = this[main];
+      if (Array.isArray(defaultValue) && Array.isArray(thisValue)) {
+        return defaultValue.every(
+          (value: any, i: number) => value === thisValue[i]
         );
       } else {
-        return proto[main] === this[main];
+        return defaultValue === thisValue;
       }
     } else {
       return false;
@@ -259,8 +259,13 @@ export abstract class AbstractBaseFilter<T> {
     }
   }
 
-  abstract applyTo2d(options: T2DPipelineState): void;
+  applyTo2d(options: T2DPipelineState): void {}
 
+  /**
+   * Returns a string that represent the current selected shader code for the filter.
+   * Used to force recompilation when parameters change or to retrieve the shader from cache
+   * @type string
+   **/
   getCacheKey() {
     return this.type;
   }
@@ -346,10 +351,10 @@ export abstract class AbstractBaseFilter<T> {
    * @param {WebGLRenderingContext} gl The canvas context used to compile the shader program.
    * @param {Object} uniformLocations A map of shader uniform names to their locations.
    */
-  abstract sendUniformData(
+  sendUniformData(
     gl: WebGLRenderingContext,
     uniformLocations: TWebGLUniformLocationMap
-  ): void;
+  ): void {}
 
   /**
    * If needed by a 2d filter, this functions can create an helper canvas to be used
@@ -384,30 +389,8 @@ export abstract class AbstractBaseFilter<T> {
     // delegate, not alias
     return this.toObject();
   }
-}
 
-export abstract class BaseFilter extends AbstractBaseFilter<string> {
-  getFragmentSource() {
-    return this.fragmentSource;
+  static async fromObject({ type, ...options }: any) {
+    return new this(options);
   }
 }
-
-Object.assign(AbstractBaseFilter.prototype, {
-  vertexSource: `
-    attribute vec2 aPosition;
-    varying vec2 vTexCoord;
-    void main() {
-      vTexCoord = aPosition;
-      gl_Position = vec4(aPosition * 2.0 - 1.0, 0.0, 1.0);
-    }`,
-});
-
-Object.assign(BaseFilter.prototype, {
-  fragmentSource: `
-    ${highPsourceCode};
-    varying vec2 vTexCoord;
-    uniform sampler2D uTexture;
-    void main() {
-      gl_FragColor = texture2D(uTexture, vTexCoord);
-    }`,
-});